fix: function to int

+++ Functions.cs
@@ -22,14 +22,14 @@ namespace Project_Info
                 Console.Write(myfile[i] + " ");
         }
 
-        public static double ConvertToInt(IEnumerable<byte> data)
+        public static int ConvertToInt(IEnumerable<byte> data)
         {
-            double result = 0;
+            int result = 0;
             var enumerable = data.ToList();
             for (var i = 0; i < enumerable.Count; i++)
             {
                 Console.WriteLine(enumerable[i]);
-                result = result + enumerable[i] * Math.Pow(256, i);
+                result = (int) (result + enumerable[i] * Math.Pow(256, i));
             }
             return result;
         }
